fix for PID change

+++ Contrib/DecayAnalysis/SemiLeptonicDecayAnalysis.h
@@ -125,7 +125,7 @@ private:
   /**
    *  Identidies of the leptons
    */
-  vector<unsigned int> _outgoingL;
+  vector<long> _outgoingL;
 
   /**
    *  Histograms
